fix: drop a redundant usize cast in the git panel
`unicode_width::UnicodeWidthStr::width` already returns `usize`, so the
cast is a no-op that a newer clippy flags under `unnecessary_cast`,
failing the `cargo clippy -- -D warnings` gate on an otherwise clean
tree. Behavior is unchanged.

+++ src/ui/git_panel.rs
@@ -454,7 +454,7 @@ pub(super) fn history_item(
     ]);
     // Right-align the short sha, leaving one trailing cell of margin.
     let sha_w = entry.short_sha.chars().count();
-    let used = subject_line.width() as usize;
+    let used = subject_line.width();
     let pad = content_width.saturating_sub(used + sha_w + 1).max(1);
     subject_line.spans.push(Span::raw(" ".repeat(pad)));
     subject_line.spans.push(Span::styled(
